requirements.txt
fastapi
uvicorn
twilio
google-generativeai
reportlab
python-multipart
sqlalchemy
python-dotenv
requests
psycopg2-binary
